提炼 common.h
