+++ TensorRTYoloV5WithoutNvSamples/TensorRTYoloV5WithoutNvSamples.vcxproj
@@ -148,6 +148,7 @@
     <ClCompile Include="YoloV5Model.cpp" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="common.h" />
     <ClInclude Include="windows\getopt.h" />
     <ClInclude Include="YoloV5Model.h" />
   </ItemGroup>
