+++ TensorRTYoloV5WithoutNvSamples/TensorRTYoloV5WithoutNvSamples.vcxproj.filters
@@ -32,5 +32,8 @@
     <ClInclude Include="windows\getopt.h">
       <Filter>头文件</Filter>
     </ClInclude>
+    <ClInclude Include="common.h">
+      <Filter>头文件</Filter>
+    </ClInclude>
   </ItemGroup>
 </Project>
